move methods around expiration from persistence to command
